(r) Use a shared utility task for schema validation
The templates and project infrastructure use tasks that validate various data files against their JSON schema. The
ajv-cli validator tool is used to perform the validation.

Previously the ajv-cli command for the validation operation was defined redundantly in each task. Considering the
command was reasonably simple, the resulting code duplication was acceptable. However, the command is flawed in that it
doesn't provide a reasonable mechanism for excluding folders in applications where the project file tree is searched
recursively for all occurrences of the data files.