Improve precomposed asset installation commit and PR messages
In order to facilitate asset installation, the installation instructions include precomposed commit and pull request
messages, which can be copy and pasted by the installer.

The primordial origin of this repository was a collection of reusable GitHub Actions workflows for use in the
repositories of Arduino libraries. In those projects, it is not common to make provisions for running the checks
locally. For this reason, the precomposed messages in the original installation instructions focused exclusively on the
GitHub Actions workflow. The installation instructions in this repository inherited that same focus. However, that
approach makes less sense in this context where the primary assets are typically general purpose infrastructure, which
may be used locally by the contributor in addition to its used by the companion GitHub Actions workflow.

The precomposed commit and PR messages are adjusted to more accurately communicate the true nature of the infrastructure
being installed.