Move comment to appropriate location
The "Check Prettier Formatting" workflow is configured to only run when relevant files are modified by the use of path
filters. Due to Prettier being a general purpose formatting tool with coverage for an array of languages, and due to the
fact that multiple file extensions may be used for files of a given language, the paths filter is extensive. It is based
on the code Prettier uses to identify files. In order to facilitate maintenance of the workflow, a comment was added to
document the source upon which the paths filter is based.

The paths filter must be defined redundantly for both the `push` and `pull_request` event triggers. The decision was
made to only add the comment to one instance of the code. Somehow it ended up being added to the second instance, where
the maintainer is less likely to see it. It is hereby moved to the first instance.